requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==19.7.1
Pillow==9.2.0
PyJWT==2.4.0
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
whitenoise==3.3.1
windows-curses==2.3.0
